refactor(sdk): rename ports.py to i_secret_provider.py for consistency with .NET SDK
