+++ src/sdks/python/envilder/domain/__init__.py
@@ -1,7 +1,7 @@
 from envilder.domain.envilder_options import EnvilderOptions
+from envilder.domain.i_secret_provider import ISecretProvider
 from envilder.domain.map_file_config import MapFileConfig
 from envilder.domain.parsed_map_file import ParsedMapFile
-from envilder.domain.ports import ISecretProvider
 from envilder.domain.secret_provider_type import SecretProviderType
 
 __all__ = [
