+++ tests/sdks/python/.gitignore
@@ -1,0 +1,1 @@
+test-results.xml
